Updated sample tests
