+++ azure-servicebus/examples/async_examples/example_session_send_receive_with_pool_async.py
@@ -5,7 +5,6 @@
 #--------------------------------------------------------------------------
 
 import asyncio
-import logging
 import sys
 import uuid
 import pytest
@@ -15,24 +14,6 @@ import conftest
 from azure.servicebus.aio import ServiceBusClient, Message
 from azure.servicebus.common.constants import NEXT_AVAILABLE
 from azure.servicebus.common.errors import NoActiveSession
-
-
-def get_logger(level):
-    azure_logger = logging.getLogger("azure")
-    if not azure_logger.handlers:
-        azure_logger.setLevel(level)
-        handler = logging.StreamHandler(stream=sys.stdout)
-        handler.setFormatter(logging.Formatter('%(asctime)s %(name)-12s %(levelname)-8s %(message)s'))
-        azure_logger.addHandler(handler)
-
-    uamqp_logger = logging.getLogger("uamqp")
-    if not uamqp_logger.handlers:
-        uamqp_logger.setLevel(logging.INFO)
-        uamqp_logger.addHandler(handler)
-    return azure_logger
-
-
-logger = get_logger(logging.INFO)
 
 
 async def message_processing(queue_client):
@@ -58,15 +39,14 @@ async def message_processing(queue_client):
             return
 
 
-async def sample_session_send_receive_with_pool(sb_config, queue):
+async def sample_session_send_receive_with_pool_async(sb_config, queue):
 
     concurrent_receivers = 5
     sessions = [str(uuid.uuid4()) for i in range(concurrent_receivers)]
     client = ServiceBusClient(
         service_namespace=sb_config['hostname'],
         shared_access_key_name=sb_config['key_name'],
-        shared_access_key_value=sb_config['access_key'],
-        debug=True)
+        shared_access_key_value=sb_config['access_key'])
 
     queue_client = client.get_queue(queue)
     for session_id in sessions:
@@ -76,11 +56,6 @@ async def sample_session_send_receive_with_pool(sb_config, queue):
 
     receive_sessions = [message_processing(queue_client) for _ in range(concurrent_receivers)]
     await asyncio.gather(*receive_sessions)
-
-
-@pytest.mark.asyncio
-async def test_async_sample_session_send_receive_with_pool(live_servicebus_config, session_queue):
-    await sample_session_send_receive_with_pool(live_servicebus_config, session_queue)
 
 
 if __name__ == '__main__':
