+++ azure-servicebus/examples/example_session_send_receive_batch.py
@@ -4,7 +4,6 @@
 # license information.
 #--------------------------------------------------------------------------
 
-import logging
 import sys
 import uuid
 
@@ -12,24 +11,6 @@ import conftest
 
 from azure.servicebus import ServiceBusClient
 from azure.servicebus import Message
-
-
-def get_logger(level):
-    azure_logger = logging.getLogger("azure")
-    if not azure_logger.handlers:
-        azure_logger.setLevel(level)
-        handler = logging.StreamHandler(stream=sys.stdout)
-        handler.setFormatter(logging.Formatter('%(asctime)s %(name)-12s %(levelname)-8s %(message)s'))
-        azure_logger.addHandler(handler)
-
-    uamqp_logger = logging.getLogger("uamqp")
-    if not uamqp_logger.handlers:
-        uamqp_logger.setLevel(logging.INFO)
-        uamqp_logger.addHandler(handler)
-    return azure_logger
-
-
-logger = get_logger(logging.INFO)
 
 
 def sample_session_send_receive_batch(sb_config, queue):
@@ -56,10 +37,6 @@ def sample_session_send_receive_batch(sb_config, queue):
             if str(message) == "shutdown":
                 session.set_session_state("END")
                 break
-
-
-def test_sample_session_send_receive_batch(live_servicebus_config, session_queue):
-    sample_session_send_receive_batch(live_servicebus_config, session_queue)
 
 
 if __name__ == '__main__':
